Broken checkin, but important milestone
Wanted to get code checked in even though it has runtime errors. This marks the
beginning of the game server and bot player interacting with AWS. Gameserver
will wait on messages from the queue Battleship_Registration.fifo and topic
Battleship_Registration for registration messages. The bot player (and eventually
the web player) will create a topic used for specific communication to that
particular player and pass that along to the registration module so it can
subscribe to it, confirm the handle and order and respond with ACK/NAK.