GameServer was not properly deleting after processing
The GameServer code did not delete messages after processing which resulted in
messages left in the queue that were picked up for re-processing.

Added bot.py which (at least for now) forks 3 'players' to register for the game.
The intention for this script is to create bot players for the game iteslf so
that they can be watched and potentially restarted if there is a problem.